split weekly planner into GUI and class and then created test for it
