+++ src/main/java/fuze/framework/gui/MainPage.java
@@ -10,6 +10,7 @@ import fuze.entity.location.LocationStringToCoordinate;
 import fuze.entity.weather.WeatherDay;
 import fuze.entity.weather.WeatherWeek;
 import fuze.usecases.generateoutfit.OutfitCreator;
+import fuze.usecases.planweekly.WeeklyPlannerInteractor;
 import fuze.usecases.planweekly.WeeklyPlanner;
 import fuze.usecases.trippacking.TripPlanner;
 import fuze.usecases.uploadclothing.*;
@@ -18,8 +19,10 @@ import javax.swing.*;
 import javax.swing.border.LineBorder;
 import java.awt.*;
 import java.nio.file.Paths;
-import java.util.*;
-import java.util.List;
+import java.util.Arrays;
+import java.util.LinkedHashSet;
+import java.util.Map;
+import java.util.Objects;
 
 public class MainPage extends JFrame {
 
@@ -32,6 +35,7 @@ public class MainPage extends JFrame {
     private final OutfitCreator outfitCreator = new OutfitCreator();
     private final LocationStringToCoordinate locationTranslator;
     private final String[] availableCities;
+    private final WeeklyPlannerInteractor weeklyPlannerInteractor;
 
     private JPanel outfitPanel;
 
@@ -46,6 +50,7 @@ public class MainPage extends JFrame {
         String[] cities = locationTranslator.getCities();
         this.availableCities = cities != null ? cities : new String[]{"Toronto Canada", "New York", "Vancouver"};
         this.currentLocation = availableCities[0];
+        this.weeklyPlannerInteractor = new WeeklyPlannerInteractor(wardrobeRepository, outfitCreator);
 
         initUi();
     }
@@ -155,7 +160,7 @@ public class MainPage extends JFrame {
     public void loadFromWeatherWeek(WeatherWeek week) {
         this.currentWeek = week;
         if (this.currentWeek != null) {
-            generateOutfitsForWeek(this.currentWeek);
+            weeklyPlannerInteractor.generateOutfitsForWeek(this.currentWeek);
         }
         WeatherDay today = week != null ? week.getWeatherDay(0) : null;
         String loc = week != null ? safe(week.getDefaultLocation()) : "Unknown";
@@ -253,148 +258,13 @@ public class MainPage extends JFrame {
 
     private void refreshOutfits() {
         currentWeek = new WeatherWeek(currentLocation);
-        generateOutfitsForWeek(currentWeek);
+        weeklyPlannerInteractor.generateOutfitsForWeek(currentWeek);
         loadFromWeatherWeek(currentWeek);
     }
 
     private void reloadWeatherForCurrentLocation() {
         WeatherWeek week = new WeatherWeek(currentLocation);
         loadFromWeatherWeek(week);
-    }
-
-    private void generateOutfitsForWeek(WeatherWeek week) {
-        Map<String, List<ClothingArticle>> wardrobeMap = buildWardrobeMap(wardrobeRepository.getAll());
-        if (wardrobeMap.isEmpty()) {
-            return;
-        }
-        Set<String> prevNames = new HashSet<>();
-        Outfit previousOutfit = null;
-        for (int i = 0; i < 7; i++) {
-            WeatherDay day = week.getWeatherDay(i);
-            if (day == null) continue;
-            Outfit outfit = generateOutfitWithNoRepeat(day, wardrobeMap, prevNames, previousOutfit);
-            day.setOutfit(outfit);
-            prevNames = extractNames(outfit);
-            previousOutfit = outfit;
-        }
-    }
-
-    private Outfit generateOutfitWithNoRepeat(WeatherDay day,
-                                              Map<String, List<ClothingArticle>> wardrobeMap,
-                                              Set<String> previousNames,
-                                              Outfit previousOutfit) {
-        Outfit first = outfitCreator.createOutfitForDay(day, shuffledCopy(wardrobeMap), false);
-        if (first == null) return null;
-        if (!isSameTopOrBottom(first, previousNames) && !isSameOutfit(first, previousOutfit)) {
-            return first;
-        }
-
-        for (int attempt = 0; attempt < 4; attempt++) {
-            Outfit candidate = outfitCreator.createOutfitForDay(day, shuffledCopy(wardrobeMap), false);
-            if (candidate != null && !isSameTopOrBottom(candidate, previousNames)
-                    && !isSameOutfit(candidate, previousOutfit)) {
-                return candidate;
-            }
-        }
-
-        Map<String, List<ClothingArticle>> filtered = new HashMap<>();
-        for (Map.Entry<String, List<ClothingArticle>> entry : wardrobeMap.entrySet()) {
-            List<ClothingArticle> list = new ArrayList<>();
-            for (ClothingArticle item : entry.getValue()) {
-                if (item != null && (previousNames == null || !previousNames.contains(item.getName()))) {
-                    list.add(item);
-                }
-            }
-            filtered.put(entry.getKey(), list);
-        }
-        Outfit lastTry = outfitCreator.createOutfitForDay(day, filtered, false);
-        if (lastTry != null && !isSameOutfit(lastTry, previousOutfit)) {
-            return lastTry;
-        }
-        return first;
-    }
-
-    private Map<String, List<ClothingArticle>> shuffledCopy(Map<String, List<ClothingArticle>> original) {
-        Map<String, List<ClothingArticle>> copy = new HashMap<>();
-        for (Map.Entry<String, List<ClothingArticle>> entry : original.entrySet()) {
-            List<ClothingArticle> list = new ArrayList<>(entry.getValue());
-            Collections.shuffle(list);
-            copy.put(entry.getKey(), list);
-        }
-        return copy;
-    }
-
-    private boolean isSameTopOrBottom(Outfit outfit, Set<String> previousNames) {
-        if (outfit == null || previousNames == null) return false;
-        Map<String, ClothingArticle> items = outfit.getItems();
-        if (items == null) return false;
-        ClothingArticle top = items.get("top");
-        ClothingArticle bottom = items.get("bottom");
-        boolean sameTop = top != null && top.getName() != null && previousNames.contains(top.getName());
-        boolean sameBottom = bottom != null && bottom.getName() != null && previousNames.contains(bottom.getName());
-        return sameTop && sameBottom;
-    }
-
-    private Set<String> extractNames(Outfit outfit) {
-        Set<String> names = new HashSet<>();
-        if (outfit == null || outfit.getItems() == null) return names;
-        for (ClothingArticle item : outfit.getItems().values()) {
-            if (item != null && item.getName() != null) {
-                names.add(item.getName());
-            }
-        }
-        return names;
-    }
-
-    /**
-     * Returns true if both outfits use the same named items for every slot.
-     * A null outfit never matches another outfit.
-     */
-    private boolean isSameOutfit(Outfit a, Outfit b) {
-        if (a == null || b == null) {
-            return false;
-        }
-        Map<String, ClothingArticle> itemsA = a.getItems();
-        Map<String, ClothingArticle> itemsB = b.getItems();
-        if (itemsA == null || itemsB == null) {
-            return false;
-        }
-        if (!itemsA.keySet().equals(itemsB.keySet())) {
-            return false;
-        }
-        for (String key : itemsA.keySet()) {
-            ClothingArticle articleA = itemsA.get(key);
-            ClothingArticle articleB = itemsB.get(key);
-            String nameA = articleA != null ? articleA.getName() : null;
-            String nameB = articleB != null ? articleB.getName() : null;
-            if (!Objects.equals(nameA, nameB)) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    private Map<String, List<ClothingArticle>> buildWardrobeMap(List<ClothingArticle> items) {
-        Map<String, List<ClothingArticle>> map = new HashMap<>();
-        map.put("top", new ArrayList<>());
-        map.put("bottom", new ArrayList<>());
-        map.put("outer", new ArrayList<>());
-        map.put("accessory", new ArrayList<>());
-
-        for (ClothingArticle item : items) {
-            if (item == null || item.getCategory() == null) continue;
-            String cat = item.getCategory().toLowerCase();
-            if (cat.contains("top")) {
-                map.get("top").add(item);
-            } else if (cat.contains("bottom") || cat.contains("pant") || cat.contains("bottoms")) {
-                map.get("bottom").add(item);
-            } else if (cat.contains("outer")) {
-                map.get("outer").add(item);
-            } else if (cat.contains("access")) {
-                map.get("accessory").add(item);
-            }
-        }
-        return map;
     }
 
     private void openTripPlanner() {
@@ -414,7 +284,7 @@ public class MainPage extends JFrame {
         if (currentWeek == null) {
             currentWeek = new WeatherWeek(currentLocation);
         }
-        generateOutfitsForWeek(currentWeek);
+        weeklyPlannerInteractor.generateOutfitsForWeek(currentWeek);
         planner.setWeatherWeek(currentWeek);
         planner.setVisible(true);
     }
